+++ frontend/index.html
@@ -1,49 +1,17 @@
 <!DOCTYPE html>
-<!--[if lt IE 7]>      <html class="no-js lt-ie9 lt-ie8 lt-ie7"> <![endif]-->
-<!--[if IE 7]>         <html class="no-js lt-ie9 lt-ie8"> <![endif]-->
-<!--[if IE 8]>         <html class="no-js lt-ie9"> <![endif]-->
-<!--[if gt IE 8]>      <html class="no-js"> <!--<![endif]-->
 <html>
-    <head>
-        <meta charset="utf-8">
-        <meta http-equiv="X-UA-Compatible" content="IE=edge">
-        <title>ToDaLoo</title>
-        <meta name="description" content="">
-        <meta name="viewport" content="width=device-width, initial-scale=1">
-        <link rel="stylesheet" href="style.css">
-    </head>
-    <body>
-        <!--[if lt IE 7]>
-            <p class="browsehappy">You are using an <strong>outdated</strong> browser. Please <a href="#">upgrade your browser</a> to improve your experience.</p>
-        <![endif]-->
-        
-        <div class="container" id="container-field">
-            <div class="todo" id="add-todo">
-                <div class="title">What do you want to accomplish next?</div>
-                <form id="input-form" onsubmit="addTask()">
-                    <input type="text" id="new-task" name="new-task" placeholder="Describe your new task">
-                    <input type="date" id="task-deadline" name="task-deadline">
-                    <select name="priority" id="priority" class="dropdown">
-                        <option value="Prio: High">High</option>
-                        <option value="Prio: Medium">Medium</option>
-                        <option value="Prio: Low">Low</option>
-                        <option value="Prio: No Prio">No Prio</option>
-                    </select>
-                    <input type="submit" value="Add task" id="submit" />
-                </form>
-            </div>
-
-            <!-- <div class="todo added" id="test">
-                <div class="todo-text">Test Todo</div>
-                <div class="todo-date">01-01-2023</div>
-                <div class="todo-prio">Prio</div>
-                <button class="todo-remove" id="remove">X</button>
-            </div> -->
-
-        </div>
-
-
-
-        <script type="module" src="./main.js" async defer></script>
-    </body>
+<head>
+	<link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap@5.0.1/dist/css/bootstrap.min.css" integrity="undefined" crossorigin="anonymous">
+	<script src="https://cdn.jsdelivr.net/npm/bootstrap@5.0.1/dist/js/bootstrap.min.js" integrity="undefined" crossorigin="anonymous"></script>
+	<link rel="stylesheet" href="/style.css">
+</head>
+<body>
+	<div class="container login-container">
+		<div class="col-sm-6 mt-5 col-sm-offset-3">
+				<h1><span class="fa fa-sign-in"></span>Welcome!</h1>
+                <button id="start" class="btn btn-lg btn-primary mt-3 ml-2">Start!</button>
+		</div>
+	</div>
+    <script src="./main.js" async defer></script>
+</body>
 </html>
